Test AvroAuth
