+++ build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlinVersion = '1.2.10'
+    ext.kotlinVersion = '1.2.21'
     dependencies {
         classpath 'org.junit.platform:junit-platform-gradle-plugin:1.0.1'
     }
@@ -7,7 +7,7 @@ buildscript {
 
 plugins {
     id 'war'
-    id "org.jetbrains.kotlin.jvm" version '1.2.10'
+    id "org.jetbrains.kotlin.jvm" version '1.2.21'
 }
 
 group 'org.radarcns'
