+++ src/main/kotlin/org/radarcns/gateway/filter/ManagementPortalAuthenticationFilter.kt
@@ -2,17 +2,12 @@ package org.radarcns.gateway.filter
 
 import org.radarcns.auth.authentication.TokenValidator
 import org.radarcns.auth.authorization.Permission
-import org.radarcns.auth.authorization.RadarAuthorization
 import org.radarcns.auth.config.YamlServerConfig
-import org.radarcns.auth.exception.NotAuthorizedException
 import org.radarcns.auth.exception.TokenValidationException
-import org.radarcns.auth.token.RadarToken
-import org.radarcns.gateway.util.Json
 import org.radarcns.gateway.util.Json.jsonErrorResponse
 import java.io.IOException
 import java.net.URI
 import java.net.URISyntaxException
-import java.util.*
 import javax.servlet.*
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
@@ -65,18 +60,19 @@ class ManagementPortalAuthenticationFilter : Filter {
 
         // Check if the HTTP Authorization header is present and formatted correctly
         if (authorizationHeader == null
-                || !authorizationHeader.toLowerCase(Locale.US).startsWith("bearer ")) {
+                || !authorizationHeader.startsWith(BEARER, ignoreCase = true)) {
             this.context.log("No authorization header provided in the request")
             return null
         }
 
         // Extract the token from the HTTP Authorization header
-        return authorizationHeader.substring("Bearer".length).trim { it <= ' ' }
+        return authorizationHeader.substring(BEARER.length).trim { it <= ' ' }
     }
 
     companion object {
 
         private var validator: TokenValidator? = null
+        const val BEARER = "Bearer "
 
         @Synchronized private fun getValidator(context: ServletContext): TokenValidator {
             if (validator == null) {
